% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlleleFrequencies.R
\name{getAlleleFrequencies}
\alias{getAlleleFrequencies}
\title{Compute allele frequencies.}
\usage{
getAlleleFrequencies(snp)
}
\arguments{
\item{snp}{Alleles encoded as 0, 1 or 2 as a number}
}
\value{
pred a vector of predictions
}
\description{
Compute allele frequenciesGenotypic
}
\details{
If AA, AB,and BB are the frequencies of the three genotypes at a locus with two alleles,
allele frequency is estimated as (2pp+pq) / (2N), where p is the frequency of A-allele;
q is the frequency of the B-allele in the sample; N is the number of samples.
}
\examples{
getAlleleFrequencies(1)
}
