% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_fit_rarecommon.R
\name{vb_fit_rarecommon}
\alias{vb_fit_rarecommon}
\title{Fit Variational Bayes model}
\usage{
vb_fit_rarecommon(
  y,
  genotype0,
  max_iter = 25000,
  weight.type = NULL,
  maf.filter.snp = 0.01,
  epsilon_conv = 1e-04,
  Bsigma2beta = 1,
  theta_beta = 0.1,
  theta_u = 0.1,
  verbose = TRUE,
  kernel = "Lin"
)
}
\arguments{
\item{y}{a phenotype vector of length n}

\item{genotype0}{a list of genotype matrices of the target sequence}

\item{max_iter}{maximum number of iteration}

\item{weight.type}{type of weight function}

\item{maf.filter.snp}{a filtering threshold of minor allele frequency for the isolated predictors}

\item{epsilon_conv}{a numeric value giving the interval endpoint}

\item{Bsigma2beta}{a numeric value for sigma beta}

\item{theta_beta}{probability of causal variants}

\item{theta_u}{probability of causal regions}

\item{verbose}{informative iteration messages}

\item{kernel}{kernel type for covariance matrix}
}
\description{
Take in genotype, phenotype and analyses the target sequence by using BLMM.
}
\details{
A hybrid model that includes a sparsity regression model and a LMM with multiple random effects.
The sparsity regression model part is designed to capture the strong predictive effects from isolated
predictors, whereas the LMM part is to capture the effects from a group of predictors located in
nearby genetic regions.
}
\examples{
data("data", package = "BLMM")
# choose model type: 1. "uw" for BLMM-UW; 2. "beta" for BLMM-BETA; . "wss" for BLMM-WSS;
fit <- vb_fit_rarecommon(y = y_train, genotype = data, weight.type = "wss")
}
