% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_predictive.R
\name{vb_predictive}
\alias{vb_predictive}
\title{Bayesian linear mixed model with multiple random effects}
\usage{
vb_predictive(fit, epsilon = 0.001, maf_beta = 0.5, maf_u = 0.5)
}
\arguments{
\item{fit}{\code{vb} fit from \code{vb_fit_rare_common}}

\item{epsilon}{a threshold for the increase of the variance}

\item{maf_beta}{a filtering threshold for the weight of parameter beta}

\item{maf_u}{a filtering threshold for the weight of parameter U}
}
\description{
Bayesian linear mixed model with multiple random effects
}
\examples{
fit <- vb_fit_rarecommon(y = y_train, genotype = data, weight.type = "wss")
vb_predictive(fit, epsilon, maf_beta, maf_u)
}
